Hang man 1
BUIDING A HANGMAN GAME
The game is currently being built in stages.

This stage is making sure all of the basic  functionalities of the game work.

At the moment the most functions created aren't responding when called and I am still figuring out how to use the random class on python asides that the input function works and my loops are working so that's progress.